+++ .gitignore
@@ -1,3 +1,12 @@
-pkg/
-api.key
-*-handin.tar.gz
+*.*/
+main/mr-tmp/
+mrtmp.*
+824-mrinput-*.txt
+/main/diff.out
+/mapreduce/x.txt
+/pbservice/x.txt
+/kvpaxos/x.txt
+*.so
+/main/mrcoordinator
+/main/mrsequential
+/main/mrworker
