--- src/.gitignore
@@ -1,9 +1,0 @@
-*.*/
-mrtmp.*
-824-mrinput-*.txt
-/main/diff.out
-/mapreduce/x.txt
-/pbservice/x.txt
-/kvpaxos/x.txt
-*.so
-*/.DS_Store
