+++ src/go.mod
@@ -1,0 +1,3 @@
+module 6.824
+
+go 1.15
