+++ src/main/diskvd.go
@@ -16,7 +16,7 @@ package main
 //   -r restart
 
 import "time"
-import "diskv"
+import "6.824/diskv"
 import "os"
 import "fmt"
 import "strconv"
