+++ src/main/lockd.go
@@ -12,7 +12,7 @@ package main
 // instead of a and b.
 
 import "time"
-import "lockservice"
+import "6.824/lockservice"
 import "os"
 import "fmt"
