--- src/main/mrmaster.go
@@ -1,33 +1,0 @@
-package main
-
-//
-// start the master process, which is implemented
-// in ../mr/master.go
-//
-// go run mrmaster.go pg*.txt
-//
-// Please do not change this file.
-//
-
-import (
-	"../mr"
-	"log"
-)
-import "time"
-import "os"
-import "fmt"
-
-func main() {
-	if len(os.Args) < 2 {
-		fmt.Fprintf(os.Stderr, "Usage: mrmaster inputfiles...\n")
-		os.Exit(1)
-	}
-
-	m := mr.MakeMaster(os.Args[1:], 10)
-	for m.Done() == false {
-		time.Sleep(time.Second)
-	}
-
-	time.Sleep(time.Second)
-	log.Printf("mrmaster existing...")
-}
