+++ src/main/mrsequential.go
@@ -7,7 +7,7 @@ package main
 //
 
 import "fmt"
-import "../mr"
+import "6.824/mr"
 import "plugin"
 import "os"
 import "log"
@@ -39,7 +39,7 @@ func main() {
 	for _, filename := range os.Args[2:] {
 		file, err := os.Open(filename)
 		if err != nil {
-			log.Fatalf("cannot open %v, err=%v", filename, err)
+			log.Fatalf("cannot open %v", filename)
 		}
 		content, err := ioutil.ReadAll(file)
 		if err != nil {
