+++ src/main/test-mr-many.sh
@@ -1,0 +1,23 @@
+#!/usr/bin/env bash
+
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 numTrials"
+    exit 1
+fi
+
+trap 'kill -INT -$pid; exit 1' INT
+
+# Note: because the socketID is based on the current userID,
+# ./test-mr.sh cannot be run in parallel
+runs=$1
+chmod +x test-mr.sh
+
+for i in $(seq 1 $runs); do
+    timeout -k 2s 900s ./test-mr.sh &
+    pid=$!
+    if ! wait $pid; then
+        echo '***' FAILED TESTS IN TRIAL $i
+        exit 1
+    fi
+done
+echo '***' PASSED ALL $i TESTING TRIALS
