+++ src/mrapps/indexer.go
@@ -7,7 +7,7 @@ package main
 //
 
 import "fmt"
-import "../mr"
+import "6.824/mr"
 
 import "strings"
 import "unicode"
