+++ src/mrapps/nocrash.go
@@ -6,7 +6,7 @@ package main
 // go build -buildmode=plugin nocrash.go
 //
 
-import "../mr"
+import "6.824/mr"
 import crand "crypto/rand"
 import "math/big"
 import "strings"
