+++ src/raft/persister.go
@@ -21,6 +21,12 @@ func MakePersister() *Persister {
 	return &Persister{}
 }
 
+func clone(orig []byte) []byte {
+	x := make([]byte, len(orig))
+	copy(x, orig)
+	return x
+}
+
 func (ps *Persister) Copy() *Persister {
 	ps.mu.Lock()
 	defer ps.mu.Unlock()
@@ -33,13 +39,13 @@ func (ps *Persister) Copy() *Persister {
 func (ps *Persister) SaveRaftState(state []byte) {
 	ps.mu.Lock()
 	defer ps.mu.Unlock()
-	ps.raftstate = state
+	ps.raftstate = clone(state)
 }
 
 func (ps *Persister) ReadRaftState() []byte {
 	ps.mu.Lock()
 	defer ps.mu.Unlock()
-	return ps.raftstate
+	return clone(ps.raftstate)
 }
 
 func (ps *Persister) RaftStateSize() int {
@@ -53,14 +59,14 @@ func (ps *Persister) RaftStateSize() int {
 func (ps *Persister) SaveStateAndSnapshot(state []byte, snapshot []byte) {
 	ps.mu.Lock()
 	defer ps.mu.Unlock()
-	ps.raftstate = state
-	ps.snapshot = snapshot
+	ps.raftstate = clone(state)
+	ps.snapshot = clone(snapshot)
 }
 
 func (ps *Persister) ReadSnapshot() []byte {
 	ps.mu.Lock()
 	defer ps.mu.Unlock()
-	return ps.snapshot
+	return clone(ps.snapshot)
 }
 
 func (ps *Persister) SnapshotSize() int {
